Update README.md
docs: Updated the readme. 
The link to the app on Google Play Store is added to the documentation.

The link to the app on Google Play Store is https://play.google.com/store/apps/details?id=com.soyinka.soyombo.naijaplusworldnews&hl=en.
This link will make it easier to have access to and install the apk file and run. There will be no need to recompile the code and run with Android Studio or others.